maskdetection/requirements.txt
numpy >= 1.19.1
requests == 2.24.0
Flask == 1.1.2
google-cloud-storage == 1.33.0


